more tests and allow custom alias_priority on fields
